Terminated driver-based executors if kill arrives before launch task.
`ExecutorRegisteredMessage` or `RunTaskMessage` may not be delivered
to a driver-based executor. Since these messages are not retried,
without this patch an executor never starts a task and remains idle,
ignoring kill task request. This patch ensures all driver-based
executors eventually shut down if kill task arrives before the task
has been started.

Review: https://reviews.apache.org/r/64033

+++ src/exec/exec.cpp
@@ -347,6 +347,16 @@ protected:
       return;
     }
 
+    // Shutdown if a kill task request is received before the task is launched.
+    // This can happen, for example, if `ExecutorRegisteredMessage` or
+    // `RunTaskMessage` have not been delivered.
+    if (!connected || !tasks.contains(taskId)) {
+      LOG(WARNING) << "Shutting down because kill task message has been"
+                   << " received before the task had been launched";
+      shutdown();
+      return;
+    }
+
     VLOG(1) << "Executor asked to kill task '" << taskId << "'";
 
     Stopwatch stopwatch;
